[BUGFIX] Prevent error when Shipping Description is null
When the Shipping Description is null, Multisafepay will error on creating the transaction and invoice for the order. Meaning the order will be paid, but will stay on 'Pending Payment' (and eventually be cancelled) in Multisafepay.

There are several reasons why ShippingDescription might be empty - in our case, we change shipping after order-save. But Magento does not actually have a 'hard' requirement on shipping description, so there might be many cases where shipping description is empty.